+++ src/index.scss
@@ -1,7 +1,7 @@
 @import 'components/searchPanel/searchPanel.scss';
 @import 'components/selectors/selectors.scss';
 @import 'components/bookItem/bookItem.scss';
-@import 'components/bookCard/bookCard.scss';
+@import 'containers/BookDetails/BookDetails.scss';
 @import 'components/paginationButton/paginationButton.scss';
 @import 'components/loader/loader.scss';
